Made LoginActivity an AccountAuthenticatorActivity and removed the auth fragment.
