--- app/src/main/res/layout/fragment_dialog_auth.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_gravity="center"
-    android:layout_height="400dp"
-    android:layout_width="320dp"
-    android:minHeight="400dp"
-    android:minWidth="320dp"
-    android:orientation="vertical">
-
-    <WebView android:id="@+id/web_oauth"
-        android:layout_height="match_parent"
-        android:layout_width="match_parent">
-
-    </WebView></LinearLayout>
